Change AST::interfaceId() to uint32_t

+++ libsolidity/ast/AST.cpp
@@ -198,11 +198,11 @@ vector<pair<util::FixedHash<4>, FunctionTypePointer>> const& ContractDefinition:
 	});
 }
 
-uint64_t ContractDefinition::interfaceId() const
+uint32_t ContractDefinition::interfaceId() const
 {
-	uint64_t result{0};
+	uint32_t result{0};
 	for (auto const& function: interfaceFunctionList(false))
-		result ^= util::fromBigEndian<uint64_t>(function.first.ref());
+		result ^= util::fromBigEndian<uint32_t>(function.first.ref());
 	return result;
 }
